HYPERFLEET-1274 - fix: remove pointer-to-interface and align Page to int64

+++ pkg/api/metadata_types.go
@@ -40,7 +40,7 @@ type Meta struct {
 
 // PagingMeta List Paging metadata
 type PagingMeta struct {
-	Page  int
+	Page  int64
 	Size  int64
 	Total int64
 }
